+++ example/data-persistence-tf/variables.tf
@@ -95,12 +95,16 @@ variable "subnets_tag_name" {
 
 variable "lambda_memory_sizes" {
   description = "Configurable map of memory sizes for lambdas"
-  type = map(string)
-  default = {}
+  type = map(number)
+  default = {
+    ProvisionPostgresDatabase = 1024 # data-persistence - works locally
+  }
 }
 
 variable "lambda_timeouts" {
   description = "Configurable map of timeouts for lambdas"
-  type = map(string)
-  default = {}
+  type = map(number)
+  default = {
+    ProvisionPostgresDatabase = 100 # data-persistence - works locally
+  }
 }