+++ example/rds-cluster-tf/variables.tf
@@ -80,12 +80,16 @@ variable "subnets_tag_name" {
 
 variable "lambda_memory_sizes" {
   description = "Configurable map of memory sizes for lambdas"
-  type = map(string)
-  default = {}
+  type = map(number)
+  default = {
+    ProvisionPostgresDatabase = 1024 # cumulus-rds-tf
+  }
 }
 
 variable "lambda_timeouts" {
   description = "Configurable map of timeouts for lambdas"
-  type = map(string)
-  default = {}
+  type = map(number)
+  default = {
+    ProvisionPostgresDatabase = 100 # cumulus-rds-tf
+  }
 }