+++ lambdas/db-provision-user-database/variables.tf
@@ -60,12 +60,12 @@ variable "dbRecreation" {
 
 variable "lambda_memory_sizes" {
   description = "Configurable map of memory sizes for lambdas"
-  type = map(string)
+  type = map(number)
   default = {}
 }
 
 variable "lambda_timeouts" {
   description = "Configurable map of timeouts for lambdas"
-  type = map(string)
+  type = map(number)
   default = {}
 }