+++ tf-modules/archive/api.tf
@@ -153,7 +153,7 @@ resource "aws_lambda_function" "private_api" {
   environment {
     variables = merge(local.api_env_variables, {"auth_mode"="private"})
   }
-  memory_size = 1280
+  memory_size = lookup(var.lambda_memory_sizes, "PrivateApiLambda", 1280)
   tags        = var.tags
 
   dynamic "vpc_config" {
@@ -178,7 +178,7 @@ resource "aws_lambda_function" "api" {
   environment {
     variables = merge(local.api_env_variables, {"auth_mode"="public"})
   }
-  memory_size = 1280
+  memory_size = lookup(var.lambda_memory_sizes, "ApiEndpoints", 1280)
   tags        = var.tags
 
   reserved_concurrent_executions = var.api_reserved_concurrency