+++ tf-modules/archive/variables.tf
@@ -212,13 +212,13 @@ variable "es_request_concurrency" {
 
 variable "lambda_memory_sizes" {
   description = "Configurable map of memory sizes for lambdas"
-  type = map(string)
+  type = map(number)
   default = {}
 }
 
 variable "lambda_timeouts" {
   description = "Configurable map of timeouts for lambdas"
-  type = map(string)
+  type = map(number)
   default = {}
 }