+++ tf-modules/cumulus/variables.tf
@@ -107,13 +107,13 @@ variable "elasticsearch_security_group_id" {
 }
 
 variable "lambda_memory_sizes" {
-  description = "Configurable map of memory sizes for ingest task lambdas in the form <lambda_identifier>_memory_size: <memory_size>"
+  description = "Configurable map of memory sizes for lambdas"
   type = map(string)
   default = {}
 }
 
 variable "lambda_timeouts" {
-  description = "Configurable map of timeouts for ingest task lambdas in the form <lambda_identifier>_timeout: <timeout>"
+  description = "Configurable map of timeouts for lambdas"
   type = map(string)
   default = {}
 }