+++ tf-modules/data-persistence/variables.tf
@@ -96,12 +96,12 @@ variable "db_migration_lambda_timeout" {
 
 variable "lambda_timeouts" {
   description = "Configurable map of timeouts for lambdas"
-  type = map(string)
+  type = map(number)
   default = {}
 }
 
 variable "lambda_memory_sizes" {
   description = "Configurable map of memory sizes for lambdas"
-  type = map(string)
+  type = map(number)
   default = {}
 }