+++ tf-modules/s3-replicator/main.tf
@@ -38,7 +38,7 @@ resource "aws_lambda_function" "s3_replicator" {
   role          = aws_iam_role.replicator_lambda_role.arn
   handler       = "index.handler"
   runtime       = "nodejs16.x"
-  timeout       = 300
+  timeout       = lookup(var.lambda_timeouts, "s3_replicator", 300)
 
   source_code_hash = data.archive_file.replicator_package.output_base64sha256