+++ tf-modules/tea-map-cache/variables.tf
@@ -48,12 +48,12 @@ variable "default_log_retention_days" {
 
 variable "lambda_memory_sizes" {
   description = "Configurable map of memory sizes for lambdas"
-  type = map(string)
+  type = map(number)
   default = {}
 }
 
 variable "lambda_timeouts" {
   description = "Configurable map of timeouts for lambdas"
-  type = map(string)
+  type = map(number)
   default = {}
 }